### 2025-08-01 #255 @Ematrion
RSTT
Submitting Author: Bucher David (@Ematrion)
Package Name: RSTT
One-Line Description of Package: Competition simulation tool to evaluate ranking
Repository Link (if existing): https://github.com/Ematrion/rstt

---

Package enable sport/games simulation providing different tournament format, state of the art rating system and probabilistic model. It is build around the interaction of a ranking,  used to seed competition, and the resulting games, used to update the ranking. It allows users to compared trained ranking with the simulation model, and answer question such has 'Does the ranking have a unique stationary distribution'.

It aims at enabling simulation based research in the context of competition.

We partner with communities to support peer review with an additional layer of
checks that satisfy community requirements. If your package fits into an

- [ ] Astropy: [My package adheres to Astropy community standards](https://www.pyopensci.org/software-peer-review/partners/astropy.html)

---

- Explain how and why the package falls under these categories (briefly, 1-2 sentences). For community partnerships, check also their specific guidelines as documented in the links above. Please note any areas you are unsure of:

It helps producing well defined, yet complex synthetic dataset with a precise and short syntax, 20 lines of code or less for millions of games. It is possible to reproduce others research (check tutorial_3) in the field.

Include but not limited to:
    - data scientist testing rating system for large video game title or sport leagues.
    - game developer who wants to make simple test.
    - teaching tool to illustrate ranking behaviour in an intuitive context.
    - mathematician studying Markov process and stationery distribution.

I know there is a paper on a tool call MMBench. I could not access it. It seems limited to its own component while RSTT allows user to define and integrate custom model. It does not seems to address the interplay between ranking and game generation project.  Does not seems to tune the initial ranking state.

- I developed this tool to perform my master thesis. I improved the code to make it a public tool, and I intend to make several publication based on it.

- Technically, the abstraction around which the tool are build, are general enough to be extended to recommandation system in a broad sense.

- It is labelled as Alpha version. With the review process I want to reach a 1.0.0 first official stable release.
- There are many features and functionalities I want to incorporate later.

[Conduct]: https://www.pyopensci.org/handbook/CODE_OF_CONDUCT.html
