### 2025-08-02 #255 @Ematrion
Link to MMBench: https://dl.acm.org/doi/10.1145/3539597.3573023

About the scope, I feel like it matches the section:
"Scope for packages that support analytics, statistics and modeling"

As for a statement of need: https://escholarship.org/content/qt2w32w4b5/qt2w32w4b5.pdf?t=p18038
D.Aldous, 2017 describes a neglected field. My tool covers this field and generalise the scope to arbitrary system, not just the Elo.

It Supports reproductible research:
[The tutorial 3](https://github.com/Ematrion/rstt/blob/main/tutorials/3_Research_Reproduction/3_Solutions.ipynb) perfectly reproduces original diagrams using RSTT simulation only.

If there is a consensual tools for simulation based research in the field, it would make life of both research and reviewer significantly easier and increase productivity.
