### 2025-08-03 dc11a23
Glicko2
+ ADD glicko2 inference system
+ TEST glicko2 passes example calculation as test
+ CHANGE in Glicko2Rating
