+++ rstt/ranking/inferer/glicko2.py
@@ -1,0 +1,210 @@
+from .glicko import Glicko
+from ..rating import GlickoRating, Glicko2Rating
+
+import math
+
+
+class Glicko2(Glicko):
+    def __init__(self, tau: float = 0.3, *args, **kwargs):
+        super().__init__(q=1, *args, **kwargs)
+        # --- system variable --- #
+        self.tau = tau  # reasonable choices are between 0.3 and 1.2
+
+        # --- scaling constant --- #
+        self.scaling = 173.718
+        self.base_mu = 1500
+
+        # algorithm constant
+        self.epsilon = 0.000000005
+        ''' NOTE on epsilon for _step5
+        
+        The value does not match
+        'The value e = 0.000001 is a sufficiently small choice'
+        Does not reproduce the iteration of the calculation example
+        e = 0.000000005 does produce very similar numerical values, and number of steps!
+        
+        Value of A,B fA, fB in step5 iteration
+        
+        epsilon = 0.00000001 
+        -5.626821433520073 -6.126821433520073 -0.0005355033552526004 1.999675064776017
+        -5.626821433520073 -5.626955295265377 -0.0002677516776263002 1.5228332788134666e-08
+        
+        epsilong = 0.000000001
+        -5.626821433520073 -6.126821433520073 -0.0005355033552526004 1.999675064776017
+        -5.626821433520073 -5.626955295265377 -0.0002677516776263002 1.5228332788134666e-08
+        -5.626955295265377 -5.626955287652446 1.5228332788134666e-08 -1.5227465554983055e-08
+        -5.626955295265377 -5.626955291458803 7.614166394067333e-09 -0.0005354469812191957
+        
+        self.epsilon = 0.000000005
+        -5.626821433520073 -6.126821433520073 -0.0005355033552526004 1.999675064776017
+        -5.626821433520073 -5.626955295265377 -0.0002677516776263002 1.5228332788134666e-08
+        -5.626955295265377 -5.626955287652446 1.5228332788134666e-08 -1.5227465554983055e-08
+.
+        
+        '''
+
+        '''
+            NOTE:
+                - Glicko2 behaves like Glicko with a q parameters set to 1.
+                This can be seen when comparing g(phi) of glicko2 with g(RD) in glicko
+                This can be seen when comparing v() of glicko2 with d2() of glicko
+
+                - step4, the DELTA function is similar to glicko r' value (Glcko.newRating)
+                But I could not figure a clean mathematical relationship.
+                Depending on its nature, it could justify a refactoring of Glicko
+
+            FIXME:
+                - Glicko2.expectedScore(mu, mu_j, phi_j)  '==' Glicko.expectedScore(r, rj, RDj)
+                with base e and  lc constant 1. since the base is not a parameter, we implement an E() method
+                This can be improved
+        '''
+
+    # --- equations --- #
+    # v(): -> glicko.d2 with q = 1
+    # g(phi) -> glicko.G with q = 1
+
+    def expectedScore(self, rating1: GlickoRating, rating2: GlickoRating, *args, **kwargs):
+        numerator = 1
+        denominator = 1 + math.exp(-self.G(rating2.sigma)
+                                   * (rating1.mu-rating2.mu)
+                                   )
+        return numerator / denominator
+
+    def f(self, x: float, phi2: float, delta2: float, v: float, a: float, tau: float):
+        numerator = math.exp(x) * (delta2 - phi2 - v - math.exp(x))
+        denominator = 2 * (phi2 + v + math.exp(x))**2
+        return (numerator / denominator) - ((x - a) / tau**2)
+
+    # --- algorithm --- #
+    def _step1(self, tau: float):
+        """Ranking Initialisation
+
+        Step 1. description of the algorithm describes operation on the ranking status.
[…]
