+++ rstt/ranking/inferer/__init__.py
@@ -4,6 +4,7 @@ It include impleentation of the Elo and Glicko rating system
 
 from .elo import Elo
 from .glicko import Glicko
+from .glicko2 import Glicko2
 from .playerlevel import PlayerLevel
 from .playerwinprc import PlayerWinPRC
 from .eventscoring import EventScoring
