### 2025-08-04 #255 @eliotwrobson
---
---

- Regarding the documentation, it looks like the documentation site only contains the API reference, while the checklist asks for more (including tutorials and installation instructions). I think the jupyter notebooks you have with worked examples satisfy this requirement however, so it would be great if these could be integrated into the docs site.
- At a glance, the readme looks good, but please be sure to go through the [README checklist](https://www.pyopensci.org/python-package-guide/documentation/repository-files/readme-file-best-practices.html#what-your-readme-md-file-should-contain), as this is what will be used in the official review.
- It looks like your code is annotated with typeguards (which provide runtime type checks), but I don't see a workflow that does static typechecking with mypy. You should consider adding this, along with a lint workflow (using ruff or a similar tool). You might also want to consider switching to uv over poetry, since uv is faster.
- On the [pypi page](https://pypi.org/project/rstt/), the emojis don't resolve correctly.
- It seems like this package could be a very useful tool for research (as you used it in your masters thesis already). If the goal is to use this as a research tool that people can cite, I would _strongly_ encourage you to complete the archiving steps needed for the JOSS submission, then people can cite the JOSS publication once that's out (https://www.pyopensci.org/software-peer-review/how-to/author-guide.html#journal-of-open-source-software-joss-submission)
- You mentioned wanting to incorporate additional features, but you don't have any open issues on the GitHub repo. Adding planned features as open issues makes it easier for external reviewers to understand the full planned scope of the project.

Overall, the package looks really good and you've clearly put a lot of work into it. With the items above resolved, this will make a strong submission. Once the above are all taken care of (at least to the degree the checklist requires), feel free to close this issue and we can get the review started. In the meantime, I'll start looking for an editor 👍🏽
