### 2025-08-08 #255 @Ematrion
Hi,

Thank you very much for your comments.
I've been working on your points - one being very challenging.
- regarding static typing, I played around mypy, pyrefly and pyright. In short. It is going to be very hard, or impossible to have meaningful static checker for the package. (Linting with ruff will be added as workflow).

 I designed rstt to test externally implemented system - i.e without knowing what the system requiers as update input nor what rating it uses. I designed the runtime-checker to provide understandable error feedback for the user. It uses a specific version of typeguard that define Protocols with members existence, not method signature.

I understand that a code which can not be verified staticaly can be challenging to review (or even use).

In the contribution.md, I do require that example code and notebook do not break. There is a workflow that runs the examples to ensure it.
