### 2025-08-08 #255 @eliotwrobson
@Ematrion that makes sense. The use of static typing isn't a hard requirement, but was just a suggestion if it wasn't too much effort on top of the type annotations you already had. No worries if it causes issues and isn't worth setting up.
