### 2025-08-09 8735e40
linting fixes
ADD __all__ to __init__.py files

FIX:
+ improves several minor syntax abuses
" Use of Any instead of any
- import * to explicit import
+ tied_items reimplementation with better static typing

REMOVE:
ranking/observer/utils.py -> to utils/observer.py
