+++ rstt/ranking/inferer/__init__.py
@@ -7,3 +7,12 @@ from .glicko import Glicko
 from .playerlevel import PlayerLevel
 from .playerwinprc import PlayerWinPRC
 from .eventscoring import EventScoring
+
+
+__all__ = [
+    Elo,
+    Glicko,
+    PlayerLevel,
+    PlayerWinPRC,
+    EventScoring
+]
