+++ rstt/scheduler/__init__.py
@@ -5,3 +5,14 @@ RSTT consider scheduler in competition as a large notion covering both sport tou
 
 from .tournament import Competition
 from .tournament import RoundRobin, SwissRound, RandomRound, SwissBracket, SingleEliminationBracket, DoubleEliminationBracket, Snake
+
+__all__ = [
+    Competition,
+    RoundRobin,
+    SwissRound,
+    RandomRound,
+    SwissBracket,
+    SingleEliminationBracket,
+    DoubleEliminationBracket,
+    Snake
+]
