+++ rstt/solver/__init__.py
@@ -13,3 +13,14 @@ Solver are responsible to produce a :class:`rstt.stypes.Score` and assign it to 
 
 
 from .solvers import BetterWin, BradleyTerry, LogSolver, CoinFlip, ScoreProb, WIN, LOSE, DRAW
+
+__all__ = [
+    BetterWin,
+    BradleyTerry,
+    LogSolver,
+    CoinFlip,
+    ScoreProb,
+    WIN,
+    LOSE,
+    DRAW
+]
