+++ rstt/ranking/standard/__init__.py
@@ -6,3 +6,13 @@ from .basicOS import BasicOS
 from .basicElo import BasicElo
 from .basicGlicko import BasicGlicko
 from .successRanking import SuccessRanking
+
+
+__all__ = [
+    BTRanking,
+    WinRate,
+    BasicOS,
+    BasicElo,
+    BasicGlicko,
+    SuccessRanking,
+]
