### 2025-08-09 7103c63
Readme updates
ADD: readme into sphinx
FIX: emoji for pypi
