+++ docs/requirements.txt
@@ -1,5 +1,6 @@
 sphinx-rtd-theme==1.3.0
 sphinx-autodoc-typehints>=1.2.0
+m2r2==0.3.4
 
 numpy==2.2.2
 names==0.3.0
