### 2025-08-09 ef148a4
Update requirements.txt
m2r2 and sphinx-rtd-theme dependancies conflict on  docutils
