+++ docs/tutorials.rst
@@ -1,0 +1,14 @@
+.. _tutorials:
+
+
+=========
+Tutorials
+=========
+
+
+Here is a collections of notebooks that helps learning the package functionnalities.
+
+.. nbgallery::
+    tutorials/1_Basics.md
+    tutorials/2_Integration.md
+    tutorials/3_Research_Reproduction.md
