+++ docs/tutorials/2_Integration.md
@@ -1,0 +1,467 @@
+# RSTT Tutorial 2
+## Integration
+
+In this notebook we will use the [openskill](https://openskill.me/en/stable/) rating system with RSTT.
+The goal is to wrapp model in a Ranking class to benefit from its functionnalities and fit in simulation.
+
+## 1. RSTT Ranking Design 
+
+A [Ranking](https://rstt.readthedocs.io/en/latest/rstt.ranking.html#rstt.ranking.ranking.Ranking) is a composition over inheritance design that contains:
+- A **[Standing](https://rstt.readthedocs.io/en/latest/rstt.ranking.html#rstt.ranking.standing.Standing)**: dict/list container hybrid. **Automaticaly sorts player** based on their *ranking point*
+- A **[RatingSystem](https://rstt.readthedocs.io/en/latest/rstt.html#rstt.stypes.RatingSystem)**: dict like container that **maps player with ratings**
+- An **[Inference](https://rstt.readthedocs.io/en/latest/rstt.html#rstt.stypes.Inference)**: provide a **.rate method()** to compute ratings
+- An **[Observer](https://rstt.readthedocs.io/en/latest/rstt.html#rstt.stypes.Observer)**: provide an **.handle_observations()** method that process ranking.update input
+
+Before integrating external system, lets start with a simple illustration.
+A ranking can be instanciated with its components specified. However, we recommand to represent a ranking design in its own class. It makes it more clear what parameters are intresect to the ranking design, and which are hyper-parameters.
+
+#### 1.1 Simple instanciation
+
+A ranking can be instanciated with its components specified. NOT RECOMMANDED
+
+
+```python
+from rstt import Ranking
+from rstt.ranking import KeyModel, Elo, GameByGame
+
+# Ambiguity about between core element and parameters. Is the handler a tunable parameter of the ranking?
+Ranking(name='elo', datamodel=KeyModel(default=1000), backend=Elo(k=20), handler=GameByGame())
+```
+
+
+
+
+    <rstt.ranking.ranking.Ranking at 0x1063ff350>
+
+
+
+#### 1.2 Class Design
+
+We recommand to represent a ranking design in its own class with an explicit naming. It makes it more clear what parameters are inherent to the ranking design, and which are tunable hyper-parameters for comparative studies.
+
+
+```python
+# Distinguish core design from parameters, handler is not a parameter.
+class EloGBG(Ranking):
+    def __init__(self, name: str, default_rating: float=1000, k: float=20):
+        # The standing component provided in the super() init.
+        super().__init__(name=name,
+                         datamodel=KeyModel(default=default_rating), # RatingSystem
+                         backend=Elo(k=20), # Inference
+                         handler=GameByGame()) # Observer
+```
+
+#### 1.3 Run illustration
+
+As you can see, there is not much to do and it works just fine in simulation. The RSTT built-in [BasicElo]() class code is in fact very similar.
+All ranking's functionalities are implemented at a higher level of abstraction and relies on minimal requirements from its components to work as intended.
+
+
+```python
+from rstt import Player, RoundRobin, LogSolver
+
+# our ranking design
+elo = EloGBG('elo')
+
+# players
+population = Player.create(nb=32)
+
+# games
+tournament = RoundRobin('test', elo, LogSolver())
+tournament.registration(population)
+tournament.run()
+
+# check if it works
+elo.update(games=tournament.games())
+elo.plot()
+```
+
+    ----------- elo -----------
+       0.       Veronica Clark       1219
+       1.     Stephanie Jordan       1191
+       2.           Chad Brown       1186
+       3.     Christopher Cade       1172
+       4.          Mable Baker       1143
[…]
