+++ docs/index.rst
@@ -3,14 +3,51 @@
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to RSTT's documentation!
-==============================================
+
+==================
+RSTT Documentation
+==================
+
+**Ranking Simulation Testing Tool**
+
+The package provides everything needed to simulate competition and generate synthetic match dataset.
+It contains ranking implementation (such as Elo and Glicko), popular tournament format (Single elimination bracket, round robin), many versus many game mode with automated outcome (score/result) generation methods.
+Additionaly different player model are available, including time varing strenght.
+
+It is a framework, letting user developp and intergrate with ease their own models to test.
+
+
+First Steps
+===========
+
+.. toctree::
+   :caption: First steps
+   :hidden:
+
+   :doc: `install`
+      Get RSTT
+
+   :doc: `usage`
+      A Simple example gives you a good overview of what the package has to offers.
+
+   :doc: `tutorials`
+      Selection of notebook
+
+
+
+Concepts
+========
+
+
+
+
 
 .. toctree::
    :maxdepth: 5
    :caption: Table of Contents:
 
    readme
+   tutorials
    modules
 
 Indices and tables
