+++ docs/readme.rst
@@ -1,4 +1,0 @@
-README
-======
-
-.. mdinclude:: ../README.md

[…]
