### 2025-08-10 ef780b4
Update index.rst
FIX: failing doc link
