### 2025-08-10 74860a3
sphinx doc
Update requiremnt and config for doc
