+++ docs/conf.py
@@ -23,8 +23,6 @@ extensions = ["sphinx.ext.todo",
               "sphinx.ext.viewcode",
               "sphinx.ext.autodoc",
               "m2r2",
-              "nbsphinx",
-              "sphinx_gallery.load_style",
               ]
 
 source_suffix = [".rst", ".md"]
