+++ docs/tutorials.rst
@@ -8,7 +8,11 @@ Tutorials
 
 Here is a collections of notebooks that helps learning the package functionnalities.
 
-.. nbgallery::
-    tutorials/1_Basics.md
-    tutorials/2_Integration.md
-    tutorials/3_Research_Reproduction.md
+   :doc:`tutorials/basics`
+      RSTT fundamentals, from concepts to user interfaces
+
+   :doc:`tutorials/intergration`
+      Run simulation with externally defined rating system
+
+   :doc:`tutorials/elo`
+      A Complete research redo in RSTT
