+++ docs/requirements.txt
@@ -1,8 +1,6 @@
 sphinx-rtd-theme>=3.0.0
 sphinx-autodoc-typehints>=1.2.0
 m2r2==0.3.4
-lxml.html.clean>=0.4.0
-nbsphinx==0.5.0
 
 
 numpy==2.2.2
