+++ docs/tutorials/basic.rst
@@ -1,0 +1,7 @@
+.. _basics:
+
+===========
+RSTT Basics
+===========
+
+..mdinclude:: 1_Basics.md
