+++ docs/tutorials/integration.rst
@@ -1,0 +1,7 @@
+.. _integration:
+
+========================
+Externaly defined system
+========================
+
+..mdinclude:: 2_Integration.md
