+++ docs/tutorials/elo.rst
@@ -1,0 +1,7 @@
+.. _elo:
+
+================
+Research Example
+================
+
+..mdinclude:: 3_Research_Reproduction.md
