### 2025-08-10 4f436da
sphinx fixes
