+++ docs/tutorials/integration.rst
@@ -4,4 +4,4 @@
 Externaly defined system
 ========================
 
-..mdinclude:: 2_Integration.md
+.. mdinclude:: 2_Integration.md
