### 2025-08-10 903cb51
sphinx updgrade
fix index and installation instruction
