### 2025-08-10 37f3e6b
more sphinx fixes
