+++ docs/index.rst
@@ -38,6 +38,31 @@ First Steps
    Selection of notebook
 
 
+Tutorials
+=========
+
+.. toctree::
+   :caption: Tutorials
+   :hidden:
+
+
+   Here is a collections of notebooks that helps learning the package functionnalities.
+
+   tutorials/basic
+   tutorials/integration
+   tutorials/elo
+
+
+:doc:`tutorials/basic`
+   RSTT fundamentals, from concepts to user interfaces
+
+:doc:`tutorials/integration`
+   Run simulation with externally defined rating system
+
+:doc:`tutorials/elo`
+   A Complete research redo in RSTT
+
+
 
 Concepts
 ========
