+++ docs/install.rst
@@ -44,18 +44,22 @@ Develloper
 RSTT is open source and welcomes any contribution. Once you have read the `guidlines`_ you can start working.
 
 1. Get the code
-.. code-block:: bash
 
+
+.. code-block:: bash
     git clone https://github.com/Ematrion/rstt.git
 
+and then 
+
+.. code-block:: bash
     cd rstt
 
 2. Install Dependencies with Poetry
 
-The project is build using `poetry`_.  If you also use it then simply:
+The project is build using `poetry`_.  If you also use it then simply, 
 
-    .. code-block:: bash
-        poetry install --with dev
+.. code-block:: bash
+    poetry install --with dev
 
 3. Alternative Installation
 
@@ -63,33 +67,30 @@ If you do use poetry you need to perform manually steps,
 
     1. Create a virtual environement
    
-    .. code-block:: bash
-        python -m venv .venv
+        .. code-block:: bash
+            python -m venv .venv
     
     2. activate it
 
-    On macOS/Linux
+        On macOS/Linux
 
-    .. code-block:: bash
-        source .venv/bin/activate
+        .. code-block:: bash
+            source .venv/bin/activate
     
-    On Windows
+        On Windows
 
-    .. code-block:: bash
-        .venv\Scripts\activate
+        .. code-block:: bash
+            .venv\Scripts\activate
 
     3. Install Dependencies
     
-    .. code-block:: bash
-        pip install --editable '.[dev]'
+        .. code-block:: bash
+            pip install --editable '.[dev]'
 
 4. Check your Installation
 
 Make sure everything works fine and run the tests with::
     pytest
-
-
-
 
 
 
