+++ docs/requirements.txt
@@ -1,6 +1,7 @@
 sphinx-rtd-theme>=3.0.0
 sphinx-autodoc-typehints>=1.2.0
 m2r2==0.3.4
+sphinx-copybutton>=0.5.0
 
 
 numpy==2.2.2
