+++ docs/requirements.txt
@@ -2,6 +2,7 @@ sphinx-rtd-theme>=3.0.0
 sphinx-autodoc-typehints>=1.2.0
 m2r2==0.3.4
 sphinx-copybutton>=0.5.0
+pygments>=2.0.0
 
 
 numpy==2.2.2
