### 2025-08-10 b9815af
clean installation instruction
