### 2025-08-11 626fd95
Doc migration
REMOVE: docs/requirements.txt

ADD: poetry docs group with sphinx spec

CHANGE: readthedocs.yaml to use poetry

FIX a bunch of .rst file

FIX: __ini__.__all__ to contain string not object
