+++ rstt/ranking/__init__.py
@@ -23,24 +23,24 @@ from .standard import (
 )
 
 __all__ = [
-    rating,
-    Standing,
-    Ranking,
-    KeyModel,
-    GaussianModel,
-    GameByGame,
-    BatchGame,
-    PlayerChecker,
-    NoHandling,
-    Elo,
-    Glicko,
-    PlayerLevel,
-    PlayerWinPRC,
-    EventScoring,
-    BTRanking,
-    WinRate,
-    SuccessRanking,
-    BasicElo,
-    BasicGlicko,
-    BasicOS,
+    "rating",
+    "Standing",
+    "Ranking",
+    "KeyModel",
+    "GaussianModel",
+    "GameByGame",
+    "BatchGame",
+    "PlayerChecker",
+    "NoHandling",
+    "Elo",
+    "Glicko",
+    "PlayerLevel",
+    "PlayerWinPRC",
+    "EventScoring",
+    "BTRanking",
+    "WinRate",
+    "SuccessRanking",
+    "BasicElo",
+    "BasicGlicko",
+    "BasicOS",
 ]
