+++ rstt/__init__.py
@@ -18,31 +18,31 @@ from .scheduler import (
     Snake
 )
 
-__all__ = [stypes,
-           config,
-           BasicPlayer,
-           Player,
-           GaussianPlayer,
-           Match,
-           Duel,
-           BetterWin,
-           BradleyTerry,
-           CoinFlip,
-           LogSolver,
-           Standing,
-           Ranking,
-           BTRanking,
-           BasicElo,
-           BasicGlicko,
-           BasicOS,
-           WinRate,
-           SuccessRanking,
-           Competition,
-           RoundRobin,
-           SwissRound,
-           RandomRound,
-           SwissBracket,
-           SingleEliminationBracket,
-           DoubleEliminationBracket,
-           Snake
+__all__ = ["stypes",
+           "config",
+           "BasicPlayer",
+           "Player",
+           "GaussianPlayer",
+           "Match",
+           "Duel",
+           "BetterWin",
+           "BradleyTerry",
+           "CoinFlip",
+           "LogSolver",
+           "Standing",
+           "Ranking",
+           "BTRanking",
+           "BasicElo",
+           "BasicGlicko",
+           "BasicOS",
+           "WinRate",
+           "SuccessRanking",
+           "Competition",
+           "RoundRobin",
+           "SwissRound",
+           "RandomRound",
+           "SwissBracket",
+           "SingleEliminationBracket",
+           "DoubleEliminationBracket",
+           "Snake"
            ]
