+++ rstt/scheduler/tournament/__init__.py
@@ -29,12 +29,12 @@ from .snake import Snake
 
 
 __all__ = [
-    Competition,
-    RoundRobin,
-    SwissRound,
-    RandomRound,
-    SwissBracket,
-    SingleEliminationBracket,
-    DoubleEliminationBracket,
-    Snake
+    "Competition",
+    "RoundRobin",
+    "SwissRound",
+    "RandomRound",
+    "SwissBracket",
+    "SingleEliminationBracket",
+    "DoubleEliminationBracket",
+    "Snake"
 ]
