+++ .readthedocs.yaml
@@ -1,33 +1,8 @@
 # Read the Docs configuration file
 # See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
 
-# Required
-#version: 2
 
-# Set the OS, Python version, and other tools you might need
-#build:
-#  os: ubuntu-24.04
-#  tools:
-#    python: "3.10"
-
-# Build documentation in the "docs/" directory with Sphinx
-#sphinx:
-#   configuration: docs/conf.py
-
-# Optionally, but recommended,
-# declare the Python requirements required to build your documentation
-# See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
-#python:
-#  install:
-#    - requirements: docs/requirements.txt
-
-
-# my additions
-#formats:
-#  - pdf
-
-
-
+# adapted from; https://test-builds.readthedocs.io/en/poetry/
 version: 2
 
 build:
