### 2025-08-12 1185969
improve testing
-REMOVE: some unused fixture and duplicated test

+ ADD: test for observer

+ ADD: game/games support for player_observer
