### 2025-08-13 a62faa5
Glicko2 testing
add test for glicko/glicko2 ranking
