### 2025-08-18 71fe7b3
Update basicGlicko.py
FIXES Ruff checks
