### 2025-08-22 bcee7af
Update tutorials and doc
DOCS:
+ Glicko2 and BasicGlicko2 Docstring
~ move roundrobin/swissround docstring to __init__
~ Add solver example for tutorial_2 notebook
+ examples/seedings.py as code example

BUG FIXES:
+ issue in keymodel __init_with_factory
Now limited to **kwargs only

TEST:
fix some test issues
+ test_keymodel default/template/factory init
+ add all examples scripts as part of examples.yaml workflow

REMOVE:
note folder. package remark are now available in wikis on github
