+++ docs/conf.py
@@ -14,7 +14,7 @@ sys.path.insert(0, os.path.abspath(".."))
 project = 'RSTT'
 copyright = '2025, David Bucher'
 author = 'David Bucher'
-release = '0.6.6.2'
+release = '0.6.7'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
