### 2025-08-22 #255 @Ematrion
I am almost done with the adjustment.

I added wikis to the project and under "Roadmap"  and started to share future features to implement. There still things to add but would that be good enough for reviewers?
