### 2025-08-23 #255 @eliotwrobson
@Ematrion I think that sounds good! Yes, there should be enough we can start looking for editors / reviewers. Feel free to close this issue once you open a full submission. Also, please let us know if you have anyone in mind that could serve as a good reviewer / editor.
